requirements.txt
streamlit>=1.28.0
numpy>=1.21.0
matplotlib>=3.5.0
torch>=1.9.0
pandas>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
plotly>=5.0.0
scikit-learn>=1.0.0
